requirements.txt
grpcio==1.44.0
grpcio-tools==1.44.0
mongoengine==0.24.1
protobuf==3.20.0
pymongo==4.1.0
six==1.16.0
